There was a problem in handling of multiple TCP segments in the TcpClient::receiveData() method. What was occuring was this:
1. octetsReceived was initialized to 0, as expected.
2. octerPtr = bufferPtr was executed, as expected.
3.  A while loop was then entered and the statement,
octetsReceived += recv(..,octetsPtr,...) was executed.
2. next, octetPtr += octetsReceived, was executed.

The result was that, on the first iteration of the while loop, everything
worked. For a single received IP datagram, there was o problem, however,
If another back-to-back IP datagram (TCP segment) arrived, the octetPtr
was not advanced by the numberof octets received by invocation of the
recv() function. Instead, the pointer was advanced by the number of octets
received plus the previous number of octets received (notice the +=).

The solution was to add another variable that was set to the number of
octets by the recv() function and advance octPtr by this new variable.

A nasty problem, indeed.